Gameplay

+++ Assets/Scripts/EnemyAI.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class EnemyAI : MonoBehaviour
 {
-    public int currentstate;
-    public int limit;
+    public float currentstate;
+    public float limit;
     // Start is called before the first frame update
     void Start()
     {
